+++ lib/mix/tasks/se_es_clean.ex
@@ -10,5 +10,6 @@ defmodule Mix.Tasks.Se.EsClean do
   def run(__args) do
     Mix.Task.run "app.start"
     ESClientApi.delete_indexes
+    ESClientApi.delete_aliases
   end
 end