+++ lib/mix/tasks/se_es_init.ex
@@ -10,5 +10,6 @@ defmodule Mix.Tasks.Se.EsInit do
   def run(__args) do
     Mix.Task.run "app.start"
     ESClientApi.create_indexes
+    ESClientApi.create_aliases
   end
 end